refactor(mcp): split toolkit_client.py into errors, path_safety, cli_adapter, schema_ops
As per issue #158 — split the MCP adapter into dedicated sub-modules:

- toolkit/mcp/errors.py: ToolkitClientError
- toolkit/mcp/path_safety.py: _safe_path, _load_cfg, TOOLKIT_ROOT, WORKSPACE_ROOT, TOOLKIT_PYTHON
- toolkit/mcp/cli_adapter.py: _toolkit_json, inspect_paths
- toolkit/mcp/schema_ops.py: show_schema, run_state, summary, blocker_hints, review_readiness, _sql_literal, _schema_from_parquet, _read_parquet_row_count, _exists

toolkit_client.py becomes thin facade with re-exports only.
server.py imports are unchanged (toolkit_client.py re-exports everything).

Fix in schema_ops: show_schema(raw) was reading dataset from wrong key.

371/371 tests pass. ruff clean.
