+++ toolkit/mcp/errors.py
@@ -1,0 +1,9 @@
+"""Error types for the MCP toolkit client."""
+
+from __future__ import annotations
+
+
+class ToolkitClientError(RuntimeError):
+    """Raised when the toolkit CLI fails or returns an unexpected result."""
+
+    pass
